Implement FilterResponseNorm and ThresholdedLinearUnit, c.f. https://arxiv.org/abs/1911.09737
I've found this a little tricky to train, the LR schedule has to be just right,
the original cosine schedule described in a paper proved a bit tricky to tune,
so I'm using the normal, ramp-up and rsqrt decay.

NOTE: I haven't made it work in the 'learn epsilon' setting, this is left as a
TODO for now.

Also re-write ResNet after extracting a couple of functions.

PiperOrigin-RevId: 285142874