New efficient attention implementations
The new versions clean up the code in preparation for adding non-causal and encoder-decoder variants. It also incorporates one more memory optimization; as a result, the Crime and Punishment colab code can now be modified to use 8 heads per layer instead of just 2.

Most of the code complexity is now localized to EfficientAttentionBase.forward_and_or_backward, which needs to handle batching for a number of regimes with different memory characteristics. A simpler reference implementation is provided, which will hopefully be enough for people interested in learning the high-level workings of the model without worrying about every last bit of memory optimization.

PiperOrigin-RevId: 292977851